Dev server

+++ app/Http/Controllers/PropertyController.php
@@ -5,15 +5,18 @@ namespace App\Http\Controllers;
 use App\Models\Notification;
 use Illuminate\Support\Facades\Session;
 use App\Models\Country;
+use App\Models\PaymentDuration;
 use App\Models\Property;
 use App\Models\PropertyCategory;
 use App\Models\PropertyType;
 use App\Models\State;
 use App\Models\Tenant;
 use App\Models\Unit;
+use App\Models\UnitType;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Gate;
 
 class PropertyController extends Controller
@@ -37,7 +40,14 @@ class PropertyController extends Controller
         $prop_cat = PropertyCategory::all();
         $prop_types = PropertyType::all();
         $countries = Country::all();
-        $properties = $user->properties;
+        $properties = Property::leftJoin('units', 'units.property_id', '=', 'properties.id')
+            ->leftJoin('tenant_rental_records', 'tenant_rental_records.unit_id', '=', 'units.id')
+            ->leftJoin('property_categories', 'properties.property_category_id', '=', 'property_categories.id')
+            ->select('properties.id', 'properties.property_name','property_categories.category_name', 'properties.property_address', DB::raw('COUNT(units.id) as total_units'), DB::raw('COUNT(tenant_rental_records.id) as total_tenants'))
+            ->groupBy('properties.id', 'properties.property_name', 'properties.property_address', 'property_categories.category_name')
+            ->where('properties.owner_id', $user->id)->get();
+
+        logInfo($properties);
 
         return view('admin.property.index')->with([
             'states' => $state,
@@ -48,10 +58,44 @@ class PropertyController extends Controller
         ]);
     }
 
+    public function propertyUnits($property_id)
+    {
+        $user = Auth::user();
+
+        # code...
+        $state = State::all();
+        $prop_cat = PropertyCategory::all();
+        $prop_types = PropertyType::all();
+        $countries = Country::all();
+        $properties = $user->properties;
+        $paymentDuration = PaymentDuration::all();
+        $property = Property::find($property_id);
+        $units = Unit::where(['property_id' => $property_id])->get();
+        $assignable_units = Unit::where(['property_id' => $property_id, 'status' => 0])->get();
+        $users = User::where('owner_id', $user->id)->get();
+
+        $unitstype = UnitType::all();
+
+        return view('admin.property.units')->with([
+            'states' => $state,
+            'prop_cats' => $prop_cat,
+            'prop_types' => $prop_types,
+            'countries' => $countries,
+            'property' => $property,
+            'users' => $users,
+            'units' => $units,
+            'assignable_units' => $assignable_units,
+            'paymentDuration' => $paymentDuration,
+            'unitstype' => $unitstype
+        ]);
+    }
+
     public function store(Request $request)
     {
         $request->validate([
             'logo' => 'required|mimes:jpeg,jpg,png,mime|max:3008',
+            'propcat' => 'required',
+            'proptype' => 'required',
         ]);
 
         $file = $request->file('logo');
@@ -67,17 +111,17 @@ class PropertyController extends Controller
 
         if ($storefile) {
             $property = Property::create([
-                'propcatId' => $request->propcat,
-                'proptypeId' => $request->proptype,
-                'ownerId' => $user->id,
-                'propname' => $request->propname,
-                'propaddress' => $request->address,
-                'propdesc' => $request->propdesc,
+                'property_category_id' => $request->propcat,
+                'proptype_id' => $request->proptype,
+                'owner_id' => $user->id,
+                'property_name' => $request->property_name,
+                'property_address' => $request->address,
+                'property_description' => $request->property_description,
                 'email' => $request->email,
                 'phone' => $request->tel,
-                'countryId' => $request->country,
-                'stateId' => $request->state,
-                'uploadsDir' => $filename,
+                'country_id' => $request->country,
+                'state_id' => $request->state,
+                'uploads_dir' => $filename,
             ]);
 
             // publish a notification for the user create action
@@ -141,7 +185,7 @@ class PropertyController extends Controller
             $filename = 'attached-file-' . $rand . time() . '.' . $file->getClientOriginalExtension();
             $file->storeAs('public/property/', $filename);
             $property->fill($input)->save();
-            $property->update(['uploadsDir' => $filename]);
+            $property->update(['uploads_dir' => $filename]);
         } else {
             $property->fill($input)->save();
         }
@@ -152,7 +196,7 @@ class PropertyController extends Controller
             'user_id' => $user->id,
             'owner_id' => $owner,
             'title' => "Property Data Updated",
-            'message' => "$user->name updated property details for $property->propname on TenancyPlus"
+            'message' => "$user->name updated property details for $property->property_name on TenancyPlus"
         ]);
 
         if ($property) {
@@ -164,15 +208,19 @@ class PropertyController extends Controller
     public function delete(Request $request)
     {
         $user = Auth::user();
-        $property =  Property::findOrFail($request->id);
-
-        $tenants = Tenant::where('propId', $request->id)->select('email')->get();
+        $property =  Property::findOrFail($request->property_id);
 
         if ($property) {
 
-            Unit::whereIn('propId', [$request->id])->select('id')->delete();
-            User::whereIn('email', $tenants->pluck('email'))->delete();
-            Tenant::whereIn('email', $tenants->pluck('email'))->delete();
+            $units = $property->units;
+
+            if (count($units) > 0) {
+                return response()->json(['status' => 1, 'message' => 'This property cannot be deleted because it contains units']);
+            }
+
+            Unit::whereIn('property_id', [$request->property_id])->select('id')->delete();
+            // User::whereIn('email', $tenants->pluck('email'))->delete();
+            // Tenant::whereIn('email', $tenants->pluck('email'))->delete();
 
             $property->delete();
 
@@ -180,11 +228,11 @@ class PropertyController extends Controller
             $notification = Notification::create([
                 'user_id' => $user->id,
                 'owner_id' => $user->owner_id,
-                'title' => "New Tenant Created",
-                'message' => "$user->name deleted $property->propname from TenancyPlus"
+                'title' => "Property Delete",
+                'message' => "$user->name deleted $property->property_name from TenancyPlus"
             ]);
         }
 
-        exit("Property Deleted successfully.");
+        return response()->json(['status' => 0, 'message' => 'Property Deleted successfully']);
     }
 }
